django_backend/requirements.txt
django
djangorestframework
python-dotenv
django-cors-headers
dj-database-url
psycopg2-binary